% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_variable.R
\name{add_variable}
\alias{add_variable}
\title{Title}
\usage{
add_variable(data, nomVariable, pourcentage, group)
}
\arguments{
\item{data}{the dataset}

\item{nomVariable}{the name of the variable that will be added}

\item{pourcentage}{the percentage of the group in the dataset}

\item{group}{The name of the groups in the variable}
}
\value{
data
}
\description{
Title
}
