% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_variant.R
\name{enrichment_variant}
\alias{enrichment_variant}
\title{Title}
\usage{
enrichment_variant(
  data_aggregated,
  variable,
  group,
  col_enrichment,
  group_enrichment,
  multiplicateur,
  time
)
}
\arguments{
\item{group_enrichment}{}
}
\description{
Title
}
