% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatif_risk.R
\name{relatif_risk}
\alias{relatif_risk}
\title{Title}
\usage{
relatif_risk(
  data_aggregated,
  variable,
  group,
  col_enrichment,
  group_enrichment
)
}
\arguments{
\item{group_enrichment}{}
}
\description{
Title
}
